+++ Usage_Samples/Dump/dumpsys-filtered_all.sh
@@ -1,0 +1,4 @@
+#!/system/bin/sh
+
+MODDIR=${0%/*/*/*}
+su -c "/system/bin/sh ${MODDIR}/extras/dumpsys-filtered.sh --all"