+++ Usage_Samples/Jitter/jitter-reducer_++wifi.sh
@@ -1,0 +1,4 @@
+#!/system/bin/sh
+
+MODDIR=${0%/*/*/*}
+su -c "/system/bin/sh ${MODDIR}/extras/jitter-reducer.sh ++wifi --status"