+++ Usage_Samples/Resampling/change-resampling-quality_179dB_408_99.sh
@@ -1,0 +1,4 @@
+#!/system/bin/sh
+
+MODDIR=${0%/*/*/*}
+su -c "/system/bin/sh ${MODDIR}/extras/change-resampling-quality.sh --cheat 179 408 99"