+++ Usage_Samples/USB/USB_48k_24_offload-hifi-playback_drc.sh
@@ -1,0 +1,4 @@
+#!/system/bin/sh
+
+MODDIR=${0%/*/*/*}
+su --mount-master -c "/system/bin/sh ${MODDIR}/USB_SampleRate_Changer.sh --drc --offload-hifi-playback 48k 24"